fix(telemetry): compile and ratchet fixes for the CODE-6/7/8 tests
- ScriptMemoryBudgetTest: the QJsonObject initialisers cast the byte counts to
  int (an unanswered 'bytes' arg of type quint64 is an ambiguous QJsonValue).
- TelemetryTransportTest: the empty-endpoint transport is brace-initialised (the
  parenthesised form is the most-vexing-parse warning, which is an error here).
- The Lua API version's two parser helpers (ScriptEngine::parseVersionHeader and
  ScriptEngine::isCompatibleVersion) move verbatim into src/core/ScriptApiVersion.cpp,
  the version's own translation unit. ScriptEngine.cpp is grandfathered by the
  500-line ratchet at 880 lines and may not grow (tests/file-length-gate.sh), and
  the CODE-6 call site adds to it; the move takes it to 850 and the version file
  (133 lines, unmeasured) gains the code that compares against its own number.
  Behaviour-preserving: same definitions, same declarations, no call site changes.

+++ tests/src/core/ScriptMemoryBudgetTest.cpp
@@ -172,7 +172,7 @@ private slots:
 
 		// In range: accepted, reported, and the record names the previous value.
 		const ControlResult set = registry->invoke(QStringLiteral("script.set_memory_budget"),
-			QJsonObject{{QStringLiteral("bytes"), 2 * OneMiB}});
+			QJsonObject{{QStringLiteral("bytes"), static_cast<int>(2 * OneMiB)}});
 		QVERIFY2(set.ok, qPrintable(set.errorMessage));
 		QCOMPARE(set.result.value(QStringLiteral("memory_budget")).toInt(), int(2 * OneMiB));
 		QCOMPARE(engine->memoryBudget(), 2 * OneMiB);
@@ -194,7 +194,7 @@ private slots:
 
 		// And the same report shows the refusal when the budget is too small.
 		QVERIFY(registry->invoke(QStringLiteral("script.set_memory_budget"),
-			QJsonObject{{QStringLiteral("bytes"), OneMiB}}).ok);
+			QJsonObject{{QStringLiteral("bytes"), static_cast<int>(OneMiB)}}).ok);
 		const ControlResult refused = registry->invoke(QStringLiteral("script.run"),
 			QJsonObject{{QStringLiteral("source"), QString::fromUtf8(MemoryHogSource)}});
 		QVERIFY2(!refused.ok, "a script over the budget ran anyway");
